pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "pandas>=3.0.2",
    "pyinstaller>=6.19.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
